+++ hutool-core/src/main/java/cn/hutool/core/bean/copier/MapToMapCopier.java
@@ -48,17 +48,16 @@ public class MapToMapCopier extends AbsCopier<Map, Map> {
 				return;
 			}
 
+			// 忽略不需要拷贝的 key,
+			if (false == copyOptions.testKeyFilter(sKeyStr)) {
+				return;
+			}
+
 			final Object targetValue = target.get(sKeyStr);
 			// 非覆盖模式下，如果目标值存在，则跳过
 			if (false == copyOptions.override && null != targetValue) {
 				return;
 			}
-
-			// 忽略不需要拷贝的 key,
-			if (true == copyOptions.testMapKeyFilter(sKey)) {
-				return;
-			}
-
 
 			// 获取目标值真实类型并转换源值
 			final Type[] typeArguments = TypeUtil.getTypeArguments(this.targetType);