+++ hutool-core/src/main/java/cn/hutool/core/bean/BeanPath.java
@@ -139,16 +139,18 @@ public class BeanPath implements Serializable {
 	 * @param bean         Bean、Map或List
 	 * @param patternParts 表达式块列表
 	 * @param value        值
+	 * @return 值
 	 */
-	private void set(Object bean, List<String> patternParts, boolean nextNumberPart, Object value) {
+	private Object set(Object bean, List<String> patternParts, boolean nextNumberPart, Object value) {
 		Object subBean = this.get(patternParts, bean, true);
 		if (null == subBean) {
 			final List<String> parentParts = getParentParts(patternParts);
 			this.set(bean, parentParts, lastIsNumber(parentParts), nextNumberPart ? new ArrayList<>() : new HashMap<>());
-			//set中有可能做过转换，因此此处重新获取bean
-			subBean = this.get(patternParts, bean, true);
+			// 避免get方法的重复调用
+			subBean = this.set(bean, parentParts, lastIsNumber(parentParts), nextNumberPart ? new ArrayList<>() : new HashMap<>());
 		}
 		BeanUtil.setFieldValue(subBean, patternParts.get(patternParts.size() - 1), value);
+		return value;
 	}
 
 	/**